swizzle/requirements.txt
ansible
jinja2
jsonpath-ng
pyyaml
stevedore
